+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/request/ShareLinkRequest.java
@@ -4,6 +4,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Data;
 import pl.cyfronet.s4e.controller.validation.Base64;
+import pl.cyfronet.s4e.controller.validation.ContentType;
+import pl.cyfronet.s4e.controller.validation.ImageDimensions;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
@@ -21,6 +23,8 @@ public class ShareLinkRequest {
 
     @NotEmpty
     @Base64
+    @ContentType(pattern = "image/(jpeg|png|gif)")
+    @ImageDimensions(maxWidth = 500, maxHeight = 500)
     @Schema(required = true, format = "base64")
     private String thumbnail;